Update hotelres.cpp
I change the code for like more like an app type code and more specified options for information like "in day 1, I want breakfast no dinner. in day2, I want just dinner." and this code checks is the hotel room is available so for an hotel this is a real deal. I hope this code help you more than the other one! Glad if I helped. :)